requirements.txt
pandas
pytz
pyTelegramBotAPI
python-dotenv
psycopg2-binary
gigachat
Flask
sqlalchemy
python-decouple
boto3